+++ src/PropTypes.js
@@ -343,3 +343,4 @@ function PropTypes () {
 
 	return propTypesObj;
 }
+