+++ src/request.js
@@ -231,3 +231,4 @@ function request () {
 
 	return Request;
 }
+