+++ src/router.js
@@ -202,3 +202,4 @@ function Router (routes, address, initialiser) {
 
 	return api;
 }
+